pyproject.toml
[tool.poetry]
name = "dashboard-demo"
version = "0.1.0"
description = "demo of using streamlit to create a dashboard"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
google-auth = "^2.28.1"
google-cloud = "^0.34.0"
google-cloud-bigquery = "*"
google-cloud-storage = "*"
pandas = "^2.2.1"
streamlit = "^1.31.1"
plotly = "^5.19.0"
numpy = "^1.26.4"
db-dtypes = "^1.2.0"
black = "^24.2.0"
streamlit-shadcn-ui = "^0.1.18"
jupyter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
good-names-rgxs="^[_a-z][_a-z0-9]?$"
